pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 99
preview = true
target-version = ["py312"]

[tool.flake8]
exclude = [
  "./.mypy_cache/*,",
  "./.venv/*",
  ".DS_Store",
  ".git",
  ".tox",
  "__pycache__"
]
ignore = ["E203", "E501", "E704", "W503"]
max-line-length = 99
plugins = ["flake8-pyproject"]

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
follow_imports_for_stubs = true
ignore_missing_imports = true
local_partial_types = true
namespace_packages = true
pretty = true
show_error_context = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.poetry]
authors = ["nikitazulikov <zhulikovnikita884@gmail.com>"]
description = ""
name = "mvs-group"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.13.3"
asyncpg = "^0.29.0"
black = "^24.8.0"
fastapi = "^0.115.0"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
httpx = "^0.27.2"
isort = "^5.13.2"
mypy = "^1.11.2"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.5.2"
pylint = "^3.3.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
python = "^3.12"
python-multipart = "^0.0.12"
redis = "^5.1.0"
sqlalchemy = "^2.0.35"
toml-sort = "^0.23.1"
uvicorn = {extras = ["standard"], version = "^0.31.0"}

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
